make formatPeriodLabel shared
