+++ apps/main/src/app/[locale]/(performance)/energy/energy-chart.tsx
@@ -6,37 +6,8 @@ import type {
 } from "@/data/progress/get-energy-history";
 import { PeriodNavigation } from "../_components/period-navigation";
 import { PeriodTabs } from "../_components/period-tabs";
+import { formatPeriodLabel } from "../_utils";
 import { EnergyChartClient } from "./energy-chart-client";
-
-function formatPeriodLabel(
-  periodStart: Date,
-  periodEnd: Date,
-  period: EnergyPeriod,
-  locale: string,
-): string {
-  if (period === "month") {
-    return new Intl.DateTimeFormat(locale, {
-      month: "long",
-      year: "numeric",
-    }).format(periodStart);
-  }
-
-  if (period === "6months") {
-    const startMonth = new Intl.DateTimeFormat(locale, {
-      month: "short",
-    }).format(periodStart);
-
-    const endMonth = new Intl.DateTimeFormat(locale, { month: "short" }).format(
-      periodEnd,
-    );
-
-    const year = periodStart.getFullYear();
-
-    return `${startMonth} - ${endMonth} ${year}`;
-  }
-
-  return String(periodStart.getFullYear());
-}
 
 export async function EnergyChart({
   average,
